+++ toolium/driver_wrapper.py
@@ -24,8 +24,8 @@ import screeninfo
 from toolium.config_driver import ConfigDriver
 from toolium.config_parser import ExtendedConfigParser
 from toolium.driver_wrappers_pool import DriverWrappersPool
-from toolium.utils import Utils
-from toolium.format_utils import get_valid_filename
+from toolium.utils.driver_utils import Utils
+from toolium.utils.path_utils import get_valid_filename
 
 
 class DriverWrapper(object):
@@ -33,7 +33,7 @@ class DriverWrapper(object):
 
     :type driver: selenium.webdriver.remote.webdriver.WebDriver or appium.webdriver.webdriver.WebDriver
     :type config: toolium.config_parser.ExtendedConfigParser or configparser.ConfigParser
-    :type utils: toolium.utils.Utils
+    :type utils: toolium.utils.driver_utils.Utils
     :type app_strings: dict
     :type session_id: str
     :type remote_node: str
@@ -122,6 +122,7 @@ class DriverWrapper(object):
             # Initialize the config object
             self.config = ExtendedConfigParser.get_config_from_file(prop_filenames)
             self.config_properties_filenames = prop_filenames
+            self.update_magic_config_names()
 
         # Override properties with system properties
         self.config.update_properties(os.environ)
@@ -129,6 +130,20 @@ class DriverWrapper(object):
         # Override properties with behave userdata properties
         if behave_properties:
             self.config.update_properties(behave_properties)
+
+    def update_magic_config_names(self):
+        """Replace '___' for ':' in options names as a workaround of a configparser limitation
+        To set a config property with : in name
+            goog:loggingPrefs = "{'performance': 'ALL', 'browser': 'ALL', 'driver': 'ALL'}"
+        Configure properties.cfg with:
+            goog___loggingPrefs: {'performance': 'ALL', 'browser': 'ALL', 'driver': 'ALL'}
+        """
+        for section in self.config.sections():
+            for option in self.config.options(section):
+                if '___' in option:
+                    option_value = self.config.get(section, option)
+                    self.config.set(section, option.replace('___', ':'), option_value)
+                    self.config.remove_option(section, option)
 
     def configure_visual_baseline(self):
         """Configure baseline directory"""